Made some changes to auto restart sevices when needed

+++ scripts/install_munge.sh
@@ -1,15 +1,21 @@
 # Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 # SPDX-License-Identifier: MIT-0
-sudo apt install -y munge libssh2-1-dev man2html sshpass -y -q
+sudo apt install -y munge libssh2-1-dev man2html sshpass -q
 
-# Try to copy down the file
-aws s3 cp s3://$CLUSTER_CONFIG_BUCKET/munge.key /etc/munge/munge.key
+# Try to copy down the file if it exists
+aws s3api head-object --bucket $CLUSTER_CONFIG_BUCKET --key munge.key || NOT_EXIST=true
+if [ ! $NOT_EXIST ]; then
+  aws s3 cp s3://$CLUSTER_CONFIG_BUCKET/munge.key /etc/munge/munge.key
+fi
 
 
 # If the file doesn't exist, then we need to create it
 if [[ ! -e /etc/munge/munge.key ]]; then
     dd if=/dev/urandom bs=1 count=1024 > /etc/munge/munge.key # From here: https://docs.01.org/clearlinux/latest/tutorials/hpc.html
-    # Copy to S3 so we can put on other instances
+fi
+
+# Copy to S3 so we can put on other instances
+if [ $NOT_EXIST ]; then
     aws s3 cp /etc/munge/munge.key s3://$CLUSTER_CONFIG_BUCKET/munge.key
 fi
